getting batching working again
